Support verity based rootfs
