+++ basefiles/journald.conf
@@ -1,8 +1,7 @@
 [Journal]
-Storage=persistent
-SystemMaxUse=1G
-SystemKeepFree=2G
-SystemMaxFileSize=100M
+SystemMaxUse=128M
+SystemKeepFree=1G
+SystemMaxFileSize=10M
 SystemMaxFiles=10
 RuntimeMaxUse=0
-ReadKMsg=yes
+ReadKMsg=no
