+++ guest-agent/dstack.toml
@@ -12,6 +12,7 @@ keys_file = "/dstack/.host-shared/.appkeys.json"
 compose_file = "/dstack/.host-shared/app-compose.json"
 public_logs = true
 public_sysinfo = true
+data_disks = ["/"]
 
 [default.core.simulator]
 enabled = false
